Support QT_SCALE_FACTOR (issue #575). More effects ported.
Scale down controls when the application window is smaller than the
threshold.
Now AkUnit oveys QT_SCALE_FACTOR.
AkTheme.controlScale can be set globally.
Renamed plugin from PrimariesColors to Saturated.
Added Zoom plugin.
Ported video effects: OilPaint, Otsu, Photocopy, Pixelate, Quark,
Radioactive, Ripple.

+++ libAvKys/Plugins/Photocopy/src/photocopyelement.h
@@ -27,12 +27,12 @@ class PhotocopyElementPrivate;
 class PhotocopyElement: public AkElement
 {
     Q_OBJECT
-    Q_PROPERTY(qreal brightness
+    Q_PROPERTY(int brightness
                READ brightness
                WRITE setBrightness
                RESET resetBrightness
                NOTIFY brightnessChanged)
-    Q_PROPERTY(qreal contrast
+    Q_PROPERTY(int contrast
                READ contrast
                WRITE setContrast
                RESET resetContrast
@@ -42,8 +42,8 @@ class PhotocopyElement: public AkElement
         PhotocopyElement();
         ~PhotocopyElement();
 
-        Q_INVOKABLE qreal brightness() const;
-        Q_INVOKABLE qreal contrast() const;
+        Q_INVOKABLE int brightness() const;
+        Q_INVOKABLE int contrast() const;
 
     private:
         PhotocopyElementPrivate *d;
@@ -55,12 +55,12 @@ class PhotocopyElement: public AkElement
         AkPacket iVideoStream(const AkVideoPacket &packet);
 
     signals:
-        void brightnessChanged(qreal brightness);
-        void contrastChanged(qreal contrast);
+        void brightnessChanged(int brightness);
+        void contrastChanged(int contrast);
 
     public slots:
-        void setBrightness(qreal brightness);
-        void setContrast(qreal contrast);
+        void setBrightness(int brightness);
+        void setContrast(int contrast);
         void resetBrightness();
         void resetContrast();
 };
